backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
marshmallow==3.21.2
oracledb>=2.0.0
python-dotenv==1.0.1
requests==2.31.0
